LU-18203 build: check OS type before using dpkg
Bright cluster manager by default installs dpkg
on it's centos/rhel installation - presumably to
allow provisioning debian nodes in the cluster,
so dpkg is in the path and can't be removed.

This patch fixes LB_USES_DPKG to check OS type
before checking if dpkg is installed.

This is a port of
Lustre-change: https://review.whamcloud.com/48616
Lustre-commit: 9c913c27e765dfd3a1a1aad47126164574508a13
Ex-bug-id: EX-5975

Test-Parameters: trivial clientdistro=ubuntu2404 testlist=runtests

Change-Id: Idc9f6edc91f9c89b40f259421b088287e08bfe9c
Signed-off-by: Jian Yu <yujian@whamcloud.com>